requirements.txt
regex==2022.6.2
pyscopg2==2.9.3
orjson==3.7.2
fastapi==0.78.0
gunicorn==20.1.0
python-Levenshtein==0.12.2
unidecode==1.3.4